Add a scan input directly on the index page
If we don't need the user to specify what kind of scan they're taking,
why make them click the header dropdown at all? I left the dropdown
input present so that it's still easy to create a new scan when you're
looking at an existing one, but this should speed things up when someone
grabs a scan and navigates to eigie.com.